fixes is api and UI for delete listing

+++ api/controllers/listing.controller.js
@@ -1,4 +1,5 @@
 import Listing from "../models/listing.model.js";
+import { errorHandler } from "../utils/error.js";
 
 export const createListing = async (req, res, next) => {
   try {
@@ -9,10 +10,16 @@ export const createListing = async (req, res, next) => {
   }
 };
 export const deleteThisListing = async (req, res, next) => {
+  const listing = await Listing.findById(req.params.id);
+  if (!listing) {
+    return next(errorHandler("Listing not found!"));
+  }
+  if (req.user.id !== listing.userRef) {
+    return next(errorHandler(401, "You can only delete your own listing!"));
+  }
   try {
-    console.log("heeeeeee");
     await Listing.findByIdAndDelete(req.params.id);
-    return res.status(201).json("Listing Deleted Successfully");
+    return res.status(200).json("Listing Deleted Successfully");
   } catch (error) {
     next(error);
   }
